requirements.txt
streamlit>=1.30.0
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.2
python-dotenv>=1.0.1

langchain>=0.1.14
langchain-openai>=0.1.3
langchain-community>=0.0.32
openai>=1.22.0